cookie added, logging cart cookie

+++ merch/models.py
@@ -5,7 +5,7 @@ import uuid
 class Merch(models.Model):
     title = models.CharField(max_length=100)
     description = models.TextField(null=True, blank=True)
-    price = models.IntegerField()
+    price = models.DecimalField(max_digits=7, decimal_places=2)
     quantity = models.IntegerField(default=0)
     image_1 = models.ImageField(null=True, blank=True)
     image_2 = models.ImageField(null=True, blank=True)
